requirements.txt
streamlit
folium
streamlit-folium
shapely
Pillow
google-genai

